fix: disable firefox cache to test a potential root cause for the build failures
The builds have worked briefly after changes that invalidate this cache, then failed again.

While running some of these builds interactively to investigate this issue, I noticed they
got auto-updated, I think this changed recently, and it's what's causing the builds to
fail when caching the firefox builds, because each downloaded build will work with
Marionette only once. What might happen is:

- On the first buiild FF will start to download an update autommtically
- We cached this FF installation
- On the next build, Marionette won't be abe to connect to FF

The reason why the second launch could be that the first run of FF was too short for
the download to complete (I would say unlikely, I'm sure the devs account for the
possibility of an incomplete download of the update) or maybe the update process
needs to run during startup, preventing Marionette from connecting.

If disabling the cache fixes the builds, it's likely the auto-update is the issue.